Cleaned up, modularized and added tooltip.
